Properly separate the 2 flags pipelineFailed / restartPipeline (...)
* The previous commit broke the usage of the _restartPipeline flag
which previously was used to trigger a restart in case of EOS. (used for
the 'loop' option I think).
I put back the code that was erroneously removed.

* Restarting the pipeline every 2 seconds of no-signal will only start if the
gstreamer pipeline state changes to FAILURE. (marked by _pipelineFailed
flag)

* The _restartPipeline is used to trigger pipeline restarts from outside
of the alignToFps thread (e.g from the busCallback in case of EOS)

* Added getPipelineState utility method