[bug fix][gstreamer] customer timestamp receiver thread not starting
Since the option to configure port 0 was added to not receive customer
timestamp data. The thread did not start because it was initialized
before parsing the yaml, therefore never actually starting the receiver
thread.
(the yaml is only parsed when adding cameras because the thread starts
only when trigger camera is added which happens after camera's manager
initialization)

Moved receiver thread initializing from constructor to add cameras (when
adding the trigger camera)